refactor: split types into index/search/cast (5c)
Search API types move to types/search.ts; the hand-written Cast SDK
surface moves out of useCast.ts into types/cast.ts; WhatsAppChannel
moves from config into types/index.ts.
